Add <Return> keybinding to button class
The command binding for a button invokes upon 1) a mouse click and
release inside the button or 2) a spacebar click and release. But I
want the button to respond to <Return> as well. To do this, I add the
line

 self.bind("<Return>", lambda e: self.clicked())

The lambda function drops the event object *.bind passes. This is
necessary since command bindings do not take arguments.