Wrap top level ttk objects in classes
I am testing an overall OOP organization for the GUI. The top level is
the MyApp class, which initializes the root. Children of the root are
all classes inheriting from ttk classes. For example MyButton inherits
from Button. This allows the overall norms of the program to be uniform
and simplifies the organization (and readability) of the main class. It
allows the main MyApp class to only handle the higher level organization
of the app, without the clutter of lowerlevel details. I have made My*
classes for Button, Frame and Notebook, so far.