Add Labelframe nested in Panedwindow in Notebook
There are some seemingly weird issues with width and height.

I have three Labelframe objects (nf1, fn2, nf3), in a vertically
oriented Panedwindow. When I don't set the height for any window, nf3
completely fills the available space. Basically, unless a height is set,
the last fills everything. If a height is set, it is treated as the
minimum.