Add resize to root
Be default the widgets do not reaact when the window is resized. By
setting the 'weight' parameter in a grid-containing widget, resizing can
be toggled. You can set a weight for each row and column, for example

a.columnconfigure(0, weight=0)
a.columnconfigure(1, weight=1)

This would result in column 1 stretching on resize but column 0
remaining constant.

In this commit, I have only configured the root frame. But the child
frames do not react to resize. Their content is constant while the root
frame shifts. This is particularly ugly for pane5, the canvas. This
canvas contains a floating widget which I can drag. But when the root is
resized, the widget can still only be moved to the constant end of the
canvas.